Fixes https://github.com/lspector/Clojush/issues/225
When going through the list of genetic operators, we now check when
there's just one operator left and choose that. This avoids the
rare but occassional situation where we'd get a `NullPointerException`
because we'd run through the whole list of operators without the sum
of the probabilities getting past the random number.