Move atom for tracking age of error vectors to `individual`
I'm not 100% sure this is right, but I decided to move the
`first-gen-for-errors` atom to `individual.clj`, and ultimately
put most of the SAMPS logic in `individual`. That makes sense
to me in the moment, but I'm not sure it's super consistent with
where things like this have been going in the past, so it may not
be the best choice.