Implement `semantics-age-meta-error`
We now have support for the age of the error vector as
a meta error. This required adding the current generation
as an entry to the `argmaps` argument sent to
`compute-errors` so we'd be able to set the age for new
error vectors and compute the age of previously seen
vectors.